pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [{ name = "narumi", email = "toucans-cutouts0f@icloud.com" }]
requires-python = ">=3.11,<3.13"
dependencies = [
    "lxml>=5.3.0",
    "markdownify>=0.13.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.6",
    "telegraph>=2.2.0",
    "youtube-search>=2.1.2",
    "numba>=0.60.0",
    "tripplus",
    "lazyopenai[langfuse]>=0.5.0",
    "mortgage>=1.0.5",
    "markdown2>=2.5.1",
    "twse>=0.2.0",
    "rich>=13.9.4",
    "yfinance>=0.2.54",
    "openai-agents>=0.0.3",
    "duckduckgo-search>=7.5.1",
    "aiocache[redis]>=0.12.3",
    "kabigon>=0.5.1",
    "uv>=0.6.12",
    "typer>=0.15.2",
    "logfire>=3.12.0",
    "loguru>=0.7.3",
]

[project.scripts]
bot = "bot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "toml>=0.10.2",
    "pip>=24.2",
    "types-cachetools>=5.5.0.20240820",
    "types-pytz>=2024.2.0.20241003",
    "types-redis>=4.6.0.20241004",
    "types-pyyaml>=6.0.12.20240917",
    "types-aiofiles>=24.1.0.20241221",
]
mlx = ["mlx-whisper>=0.4.1"]

[tool.uv.sources]
tripplus = { git = "https://github.com/narumiruna/tripplus.git" }

[tool.ruff]
exclude = ["build"]
line-length = 120

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
